APi Cache for(sqlite and contentprovider)followers recyclerview

+++ app/src/main/java/com/example/tony/simpletwitter/models/Follower.java
@@ -69,7 +69,7 @@ public class Follower implements Parcelable{
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(long id) {
         this.id = id;
     }
